requirements.txt
opencv-python
numpy
mss
plyer
pystray
pillow
winotify